refactor backup handler

+++ writers_test.go
@@ -112,7 +112,7 @@ func (suite *writersTestSuite) TestTokenStatsWriter() {
 	mockStats := newMockStatsSetterToken(suite.T())
 	mockStats.EXPECT().AddUDFs(uint32(1))
 	mockStats.EXPECT().AddSIndexes(uint32(1))
-	mockStats.EXPECT().AddTotalBytesWritten(uint64(1))
+	mockStats.EXPECT().AddTotalBytesWritten(1)
 
 	writer := newWriterWithTokenStats(mockWriter, mockStats, slog.Default())
 	suite.NotNil(writer)
